+++ src/main/java/com/norialertapp/repository/VariantRepo.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * Created by katherine_celeste on 9/30/16.
  */
 @Repository
-public interface VariantRepo extends JpaRepository<Variant, Integer> {
+public interface VariantRepo extends JpaRepository<Variant, Long> {
 }
